remove bashrc, moved to some global profile file
if I ssh to remote, depending on (is interactive),(is login),(and maybe something)
bash decides whether it wants to evaluate this file.
I want the code to run always (although its a mess at the moment),
the interactive checks are already in the code.
So i can see whats happening in the code (only one file to open),and i dont have to remember
the behaviour.